fpl-agent/requirements.txt
langchain
langchain-google-genai
requests
python-dotenv
ratelimit